requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.8.5
djangorestframework==3.2.2
gunicorn==19.4.5
kombu==3.0.35
MySQL-python==1.2.5
python-dateutil==2.5.2
redis==2.10.5
requests==2.9.1
six==1.10.0
wheel==0.24.0
